Refactor pkg/svat/kube to pkg/uti/kube
